requirements.txt
python-fedora==0.10.0

